Added health indication

+++ Assets/Scripts/Health.cs
@@ -2,9 +2,9 @@ using UnityEngine;
 using System.Collections;
 
 public class Health : MonoBehaviour {
-	
-	public int healthLevel;
-	public int maxHealth;
+
+	public float healthLevel;
+	public float maxHealth;
 	
 	public void increaseHealth (int delta) {
 		healthLevel = Mathf.Min(healthLevel + delta, maxHealth);
@@ -12,7 +12,7 @@ public class Health : MonoBehaviour {
 	
 	public void decreaseHealth (int delta) {
 		healthLevel -= delta;	
-		if (healthLevel <= 0)
+		if (healthLevel <= 0f)
 			die();
 	}
 	
@@ -20,4 +20,5 @@ public class Health : MonoBehaviour {
 		//Destruction/death animation/time travel.
 		Destroy(this.gameObject);
 	}
+
 }
